+++ services/order/internal/repository/order.go
@@ -9,7 +9,7 @@ import (
 type OrderRepository interface {
 	CreateOrder(ctx context.Context, order domain.Order) (string, error)
 	UpdateStatusByPaymentID(ctx context.Context, yookassaPaymentID, newStatus string) error
-	GetOrderByPublicID(ctx context.Context, publicID string, userID int) (domain.Order, error)
+	GetOrderByPublicID(ctx context.Context, publicID string, userID int64) (domain.Order, error)
 	SetYookassaID(ctx context.Context, orderPublicID, yookassaID string) error
-	GetOrdersByUserID(ctx context.Context, userID int) ([]domain.Order, error)
+	GetOrdersByUserID(ctx context.Context, userID int64) ([]domain.Order, error)
 }
