+++ services/payment/internal/repository/payment.go
@@ -8,14 +8,14 @@ import (
 )
 
 type PaymentRepository interface {
-	Create(ctx context.Context, method domain.PaymentMethod) (int, error)
-	Delete(ctx context.Context, cardID string, userID int) error
-	GetByUserID(ctx context.Context, userID int) ([]domain.PaymentMethod, error)
-	SetDefault(ctx context.Context, cardID string, userID int) error
+	Create(ctx context.Context, method domain.PaymentMethod) (int64, error)
+	Delete(ctx context.Context, cardID string, userID int64) error
+	GetByUserID(ctx context.Context, userID int64) ([]domain.PaymentMethod, error)
+	SetDefault(ctx context.Context, cardID string, userID int64) error
 }
 
 type PaymentCacheRepository interface {
-	SetPendingBinding(ctx context.Context, paymentID string, userID int, ttl time.Duration) error
+	SetPendingBinding(ctx context.Context, paymentID string, userID int64, ttl time.Duration) error
 	DeletePendingBinding(ctx context.Context, paymentID string) error
-	GetUserIDByPaymentID(ctx context.Context, paymentID string) (int, error)
+	GetUserIDByPaymentID(ctx context.Context, paymentID string) (int64, error)
 }
