+++ services/restaurant/internal/repository/restaurant.go
@@ -8,5 +8,5 @@ import (
 
 type RestaurantBrandRepository interface {
 	GetRestaurantBrandsList(ctx context.Context, limit, offset int) ([]domain.RestaurantBrand, error)
-	GetByID(ctx context.Context, id int) (domain.RestaurantBrand, error)
+	GetByID(ctx context.Context, id int64) (domain.RestaurantBrand, error)
 }
